+++ packages/webblocks-cms/src/Support/WebBlocks.php
@@ -10,7 +10,7 @@ final class WebBlocks
 
     public const HANDLE = 'webblocks-cms';
 
-    public const VERSION = '1.32.44';
+    public const VERSION = '1.32.45';
 
     public const UI_VERSION = 'v2.7.7';